+++ ouroboros-network/src/Ouroboros/Network/PeerSelection/Types.hs
@@ -12,7 +12,12 @@ data PeerSource = PeerSourceLocalRoot
   deriving (Eq, Ord, Show, Enum)
 
 data PeerStatus =
-       PeerCold
+       PeerReallyCold
+     -- ^ Peer is in true cold which means no connection to exists
+     -- and the outbound governor is safe to promote it.
+     | PeerCold
+     -- ^ Peer is in cold state but its connection still lingers.
+     -- I.e. it is still in progress to be fully demoted.
      | PeerWarm
      | PeerHot
   deriving (Eq, Ord, Show)